refactor: SessionManager 및 Session 클래스 리팩터링
- HashMap을 ConcurrentHashMap으로 변경하여 동시성 처리 강화
- 패키지를 org.apache.coyote에서 org.apache.catalina로 이동하여 구조 개선
