+++ tomcat/src/main/java/org/apache/catalina/session/SessionManager.java
@@ -1,12 +1,12 @@
-package org.apache.coyote.http11;
+package org.apache.catalina.session;
 
 import com.techcourse.model.User;
-import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 public class SessionManager {
 
-    private static final Map<String, Session> sessionMap = new HashMap<>();
+    private static final Map<String, Session> sessionMap = new ConcurrentHashMap<>();
 
     public void addSession(String sessionId, Session session){
         sessionMap.put(sessionId, session);
